+++ config/default.json
@@ -5,6 +5,11 @@
             "rpId": "localhost",
             "challengeSize": 32
         },
-        "origin": "http://localhost:3000"
+        "origin": "http://localhost:3000",
+        "cookie": {
+            "name": "webauthn",
+            "maxAge": 60000,
+            "httpOnly": true
+        }
     }
 }
