+++ example/app.js
@@ -28,13 +28,10 @@ app.use(cookieSession({
   maxAge: 24 * 60 * 60 * 1000
 }));
 app.use('/public', express.static(path.join(__dirname, 'public')));
+app.use(index.webAuthentication);
 
 app.use('/', indexRouter);
 app.use('/users', usersRouter);
-app.use('/attestation/options', index.attestationOptions);
-app.use('/attestation/result', index.attestationResult);
-app.use('/assertion/options', index.assertionOptions);
-app.use('/assertion/result', index.assertionResult);
 
 // catch 404 and forward to error handler
 app.use(function (req, res, next) {
