+++ lib/src/index.js
@@ -38,341 +38,86 @@ var __importDefault = (this && this.__importDefault) || function (mod) {
     return (mod && mod.__esModule) ? mod : { "default": mod };
 };
 Object.defineProperty(exports, "__esModule", { value: true });
-var util_1 = require("./util");
 var config_1 = __importDefault(require("config"));
-var base64url_1 = __importDefault(require("base64url"));
-var fido2lib = require("fido2-lib");
-var str2ab = require("string-to-arraybuffer");
-var cache = require("./cache");
 var fido2MiddlewareConfig = config_1.default.get("fido2-middlewareConfig");
-function preFormatResultReq(reqBody) {
-    if (reqBody.response.authenticatorData) {
-        reqBody.response.authenticatorData = util_1.toArrayBuffer(base64url_1.default.toBuffer(reqBody.response.authenticatorData));
-    }
-    reqBody.id = str2ab(reqBody.id);
-    reqBody.rawId = str2ab(reqBody.rawId);
-    return reqBody;
-}
-function findAuthr(credID, authenticators) {
-    for (var _i = 0, authenticators_1 = authenticators; _i < authenticators_1.length; _i++) {
-        var authr = authenticators_1[_i];
-        if (authr.credID === credID)
-            return authr;
-    }
-    throw new Error("Unknown authenticator with credID " + credID);
-}
-/**
- *
- * @param {Object} req - Express request object
- * @param {Object} res - Express response object
- * @param {Function} next - Express next middleware function
- * @returns {undefined}
- */
-function attestationOptions(req, res) {
+var attestation_1 = require("./attestation");
+var assertion_1 = require("./assertion");
+function webAuthentication(req, res, next) {
     return __awaiter(this, void 0, void 0, function () {
-        var excludeCredentials, cacheData, fido2Lib, result, options;
-        return __generator(this, function (_a) {
-            switch (_a.label) {
+        var errorMessage, _a, e_1, e_2, e_3, e_4;
+        return __generator(this, function (_b) {
+            switch (_b.label) {
                 case 0:
-                    if (!req.body || !req.body.username || !req.body.displayName) {
-                        return [2 /*return*/, res.json({
-                                status: "failed",
-                                errorMessage: "Request missing display name or username field!"
-                            })];
+                    console.log(req.url);
+                    _a = req.url;
+                    switch (_a) {
+                        case fido2MiddlewareConfig.attestationOptionsPath || "/attestation/options": return [3 /*break*/, 1];
+                        case fido2MiddlewareConfig.attestationResultPath || "/attestation/result": return [3 /*break*/, 5];
+                        case fido2MiddlewareConfig.assertionOptionsPath || "/assertion/options": return [3 /*break*/, 9];
+                        case fido2MiddlewareConfig.assertionResultPath || "/assertion/result": return [3 /*break*/, 13];
                     }
-                    if (!!fido2MiddlewareConfig.db) return [3 /*break*/, 4];
-                    return [4 /*yield*/, cache.getAsync(req.body.username)];
+                    return [3 /*break*/, 17];
                 case 1:
-                    cacheData = _a.sent();
-                    if (!(cacheData && cacheData.registered)) return [3 /*break*/, 2];
-                    excludeCredentials = [
-                        {
-                            type: "public-key",
-                            id: cacheData.authenticators[0].credID
-                        }
-                    ];
+                    _b.trys.push([1, 3, , 4]);
+                    return [4 /*yield*/, attestation_1.attestationOptions(req, res)];
+                case 2:
+                    _b.sent();
                     return [3 /*break*/, 4];
-                case 2: return [4 /*yield*/, cache.setAsync(req.body.username, {
-                        name: req.body.displayName,
-                        registerd: false,
-                        id: util_1.randomBase64URLBuffer(32),
-                        authenticators: []
-                    })];
                 case 3:
-                    _a.sent();
-                    _a.label = 4;
-                case 4:
-                    fido2Lib = new fido2lib.Fido2Lib(fido2MiddlewareConfig.fido2lib);
-                    return [4 /*yield*/, fido2Lib.attestationOptions().catch(function (err) {
-                            return res.json({
-                                status: "failed",
-                                errorMessage: err.message
-                            });
-                        })];
+                    e_1 = _b.sent();
+                    errorMessage = e_1.message;
+                    return [3 /*break*/, 4];
+                case 4: return [3 /*break*/, 17];
                 case 5:
-                    result = _a.sent();
-                    result.user.name = req.body.username;
-                    result.user.id = util_1.randomBase64URLBuffer(32);
-                    result.user.displayName = req.body.displayName;
-                    options = {
-                        rp: result.rp,
-                        user: result.user,
-                        challenge: util_1.randomBase64URLBuffer(32),
-                        pubKeyCredParams: result.pubKeyCredParams,
-                        timeout: result.timeout,
-                        attestation: result.attestation,
-                        status: "ok",
-                        errorMessage: "",
-                        extensions: req.body.extensions,
-                        authenticatorSelection: req.body.authenticatorSelection,
-                        excludeCredentials: excludeCredentials
-                    };
-                    if (req.session) {
-                        req.session.challenge = options.challenge;
-                        req.session.username = req.body.username;
-                    }
-                    return [2 /*return*/, res.json(options)];
-            }
-        });
-    });
-}
-/**
- *
- * @param {Object} req - Express request object
- * @param {Object} res - Express response object
- * @param {Function} next - Express next middleware function
- * @returns {undefined}
- */
-function attestationResult(req, res) {
-    return __awaiter(this, void 0, void 0, function () {
-        var errorMessage, isBase64Url, fido2Lib, expected, requestBody, result, authrInfo, cacheData, authenticators, obj;
-        return __generator(this, function (_a) {
-            switch (_a.label) {
-                case 0:
-                    if (!(req.body != null && util_1.isRequestBody(req.body))) {
-                        return [2 /*return*/, res.json({
-                                status: "failed",
-                                errorMessage: "Response missing one or more of id/rawId/response/type fields"
-                            })];
-                    }
-                    if (req.body.type !== "public-key") {
-                        return [2 /*return*/, res.json({
-                                status: "failed",
-                                errorMessage: "type is not public-key!"
-                            })];
-                    }
-                    try {
-                        isBase64Url = util_1.isBase64UrlEncoded(req.body.id);
-                    }
-                    catch (e) {
-                        errorMessage = e.message;
-                    }
-                    if (!isBase64Url) {
-                        return [2 /*return*/, res.json({
-                                status: "failed",
-                                errorMessage: errorMessage || "Invalid id!"
-                            })];
-                    }
-                    fido2Lib = new fido2lib.Fido2Lib();
-                    expected = {
-                        challenge: req.session ? req.session.challenge : "",
-                        origin: fido2MiddlewareConfig.origin || "localhost",
-                        factor: fido2MiddlewareConfig.factor || "either"
-                    };
-                    requestBody = preFormatResultReq(req.body);
-                    return [4 /*yield*/, fido2Lib
-                            .attestationResult(requestBody, expected)
-                            .catch(function (err) {
-                            errorMessage = err.message;
-                        })];
-                case 1:
-                    result = _a.sent();
-                    if (!result || errorMessage) {
+                    _b.trys.push([5, 7, , 8]);
+                    return [4 /*yield*/, attestation_1.attestationResult(req, res)];
+                case 6:
+                    _b.sent();
+                    return [3 /*break*/, 8];
+                case 7:
+                    e_2 = _b.sent();
+                    errorMessage = e_2.message;
+                    return [3 /*break*/, 8];
+                case 8: return [3 /*break*/, 17];
+                case 9:
+                    _b.trys.push([9, 11, , 12]);
+                    return [4 /*yield*/, assertion_1.assertionOptions(req, res)];
+                case 10:
+                    _b.sent();
+                    return [3 /*break*/, 12];
+                case 11:
+                    e_3 = _b.sent();
+                    errorMessage = e_3.message;
+                    return [3 /*break*/, 12];
+                case 12: return [3 /*break*/, 17];
+                case 13:
+                    _b.trys.push([13, 15, , 16]);
+                    return [4 /*yield*/, assertion_1.assertionResult(req, res)];
+                case 14:
+                    _b.sent();
+                    return [3 /*break*/, 16];
+                case 15:
+                    e_4 = _b.sent();
+                    errorMessage = e_4.message;
+                    return [3 /*break*/, 16];
+                case 16: return [3 /*break*/, 17];
+                case 17:
+                    if (errorMessage) {
                         return [2 /*return*/, res.json({
                                 status: "failed",
                                 errorMessage: errorMessage
                             })];
                     }
-                    authrInfo = {
-                        fmt: result.authnrData.get("fmt"),
-                        publicKey: result.authnrData.get("credentialPublicKeyPem"),
-                        counter: result.authnrData.get("counter"),
-                        credID: base64url_1.default.encode(result.authnrData.get("credId"))
-                    };
-                    if (!req.session) return [3 /*break*/, 4];
-                    return [4 /*yield*/, cache.getAsync(req.session.username)];
-                case 2:
-                    cacheData = _a.sent();
-                    authenticators = [];
-                    if (cacheData) {
-                        authenticators = cacheData.authenticators;
-                    }
-                    authenticators.push(authrInfo);
-                    obj = {
-                        authenticators: authenticators,
-                        registered: true
-                    };
-                    return [4 /*yield*/, cache.setAsync(req.session.username, obj)];
-                case 3:
-                    _a.sent();
-                    req.session.loggedIn = true;
-                    _a.label = 4;
-                case 4: return [2 /*return*/, res.json({
-                        status: "ok",
-                        errorMessage: ""
-                    })];
-            }
-        });
-    });
-}
-/**
- *
- * @param {Object} req - Express request object
- * @param {Object} res - Express response object
- * @returns {undefined}
- */
-function assertionOptions(req, res) {
-    return __awaiter(this, void 0, void 0, function () {
-        var cacheData, authenticators, allowCredentials, _i, authenticators_2, authr, serve, result, options;
-        return __generator(this, function (_a) {
-            switch (_a.label) {
-                case 0:
-                    if (!req.body || !req.body.username) {
-                        return [2 /*return*/, res.json({
-                                status: "failed",
-                                errorMessage: "Request missing username field!"
-                            })];
-                    }
-                    return [4 /*yield*/, cache.getAsync(req.body.username)];
-                case 1:
-                    cacheData = _a.sent();
-                    authenticators = cacheData.authenticators;
-                    allowCredentials = [];
-                    for (_i = 0, authenticators_2 = authenticators; _i < authenticators_2.length; _i++) {
-                        authr = authenticators_2[_i];
-                        allowCredentials.push({
-                            type: "public-key",
-                            id: authr.credID
-                        });
-                    }
-                    serve = new fido2lib.Fido2Lib();
-                    return [4 /*yield*/, serve.assertionOptions()];
-                case 2:
-                    result = _a.sent();
-                    options = {
-                        challenge: util_1.randomBase64URLBuffer(32),
-                        timeout: result.timeout,
-                        status: "ok",
-                        allowCredentials: allowCredentials,
-                        errorMessage: "",
-                        extensions: req.body.extensions,
-                        userVerification: req.body.userVerification || "preferred"
-                    };
-                    if (req.session) {
-                        req.session.challenge = options.challenge;
-                        req.session.username = req.body.username;
-                        req.session.userVerification = options.userVerification;
-                    }
-                    return [2 /*return*/, res.json(options)];
-            }
-        });
-    });
-}
-/**
- *
- * @param {Object} req - Express request object
- * @param {Object} res - Express response object
- * @param {Function} next - Express next middleware function
- * @returns {undefined}
- */
-function assertionResult(req, res) {
-    return __awaiter(this, void 0, void 0, function () {
-        var errorMessage, isBase64Url, clientData, authenticators, cacheData, authr, fido2Lib, expected, requestBody, result;
-        return __generator(this, function (_a) {
-            switch (_a.label) {
-                case 0:
-                    if (!(req.body != null && util_1.isRequestBody(req.body))) {
-                        return [2 /*return*/, res.json({
-                                status: "failed",
-                                errorMessage: "Response missing one or more of id/rawId/response/type fields"
-                            })];
-                    }
-                    if (req.body.type !== "public-key") {
-                        return [2 /*return*/, res.json({
-                                status: "failed",
-                                errorMessage: "type is not public-key!"
-                            })];
-                    }
-                    try {
-                        isBase64Url = util_1.isBase64UrlEncoded(req.body.id);
-                    }
-                    catch (e) {
-                        errorMessage = e.message;
-                    }
-                    if (!isBase64Url || errorMessage) {
-                        res.json({
-                            status: "failed",
-                            errorMessage: errorMessage || "Invalid id!"
-                        });
-                        return [2 /*return*/];
-                    }
-                    clientData = {
-                        challenge: "",
-                        type: "",
-                        origin: ""
-                    };
-                    try {
-                        clientData = JSON.parse(base64url_1.default.decode(req.body.response.clientDataJSON));
-                    }
-                    catch (e) {
-                        errorMessage = e.message;
-                    }
-                    if (!req.session) return [3 /*break*/, 2];
-                    return [4 /*yield*/, cache.getAsync(req.session.username)];
-                case 1:
-                    cacheData = _a.sent();
-                    authenticators = cacheData.authenticators;
-                    _a.label = 2;
-                case 2:
-                    authr = findAuthr(req.body.id, authenticators);
-                    fido2Lib = new fido2lib.Fido2Lib();
-                    expected = {
-                        challenge: clientData.challenge,
-                        origin: fido2MiddlewareConfig.origin || "localhost",
-                        factor: fido2MiddlewareConfig.factor || "either",
-                        publicKey: authr.publicKey,
-                        prevCounter: authr.counter,
-                        userHandle: null
-                    };
-                    requestBody = preFormatResultReq(req.body);
-                    return [4 /*yield*/, fido2Lib
-                            .assertionResult(requestBody, expected)
-                            .catch(function (err) {
-                            errorMessage = err.message;
-                        })];
-                case 3:
-                    result = _a.sent();
-                    if (!result || errorMessage) {
-                        return [2 /*return*/, res.json({
-                                status: "failed",
-                                errorMessage: errorMessage
-                            })];
-                    }
-                    if (req.session) {
-                        req.session.loggedIn = true;
-                    }
-                    return [2 /*return*/, res.json({
-                            status: "ok",
-                            errorMessage: ""
-                        })];
+                    //  } else {
+                    //    next();
+                    //  }
+                    next();
+                    return [2 /*return*/];
             }
         });
     });
 }
 module.exports = {
-    attestationOptions: attestationOptions,
-    attestationResult: attestationResult,
-    assertionOptions: assertionOptions,
-    assertionResult: assertionResult
+    webAuthentication: webAuthentication
 };
 //# sourceMappingURL=index.js.map
