+++ lib/src/util.js
@@ -5,6 +5,7 @@ var __importDefault = (this && this.__importDefault) || function (mod) {
 Object.defineProperty(exports, "__esModule", { value: true });
 var crypto_1 = __importDefault(require("crypto"));
 var base64url_1 = __importDefault(require("base64url"));
+var str2ab = require("string-to-arraybuffer");
 /**
  *
  * @param {Number} len
@@ -48,4 +49,13 @@ function isRequestBody(bodyObject) {
     return (bodyObject.id && bodyObject.rawId && bodyObject.response && bodyObject.type);
 }
 exports.isRequestBody = isRequestBody;
+function preFormatResultReq(reqBody) {
+    if (reqBody.response.authenticatorData) {
+        reqBody.response.authenticatorData = toArrayBuffer(base64url_1.default.toBuffer(reqBody.response.authenticatorData));
+    }
+    reqBody.id = str2ab(reqBody.id);
+    reqBody.rawId = str2ab(reqBody.rawId);
+    return reqBody;
+}
+exports.preFormatResultReq = preFormatResultReq;
 //# sourceMappingURL=util.js.map
