+++ lib/test/assertion.test.d.ts
@@ -1,0 +1,1 @@
+export {};
