+++ lib/test/attestation.test.js
@@ -1,0 +1,38 @@
+"use strict";
+Object.defineProperty(exports, "__esModule", { value: true });
+var mocha_1 = require("mocha");
+var index = require("../src/attestation");
+var request = require("supertest");
+var express = require("express");
+var bodyParser = require("body-parser");
+var cookieSession = require("cookie-session");
+var cookieParser = require("cookie-parser");
+var crypto = require("crypto");
+mocha_1.describe("attestationOptions()", function () {
+    var app = express();
+    app.use(bodyParser.json());
+    app.use(cookieSession({
+        name: "session",
+        keys: [crypto.randomBytes(32).toString("hex")],
+        maxAge: 24 * 60 * 60 * 1000
+    }));
+    app.use(cookieParser());
+    app.use("/", index.attestationOptions);
+    mocha_1.it("should return ok status", function (done) {
+        request(app)
+            .post("/")
+            .send({ username: "johndoe@example.com", displayName: "John Doe" })
+            .set("Accept", "application/json")
+            .expect("Content-Type", /json/)
+            .expect(200, done);
+    });
+    mocha_1.it("should return failed status", function (done) {
+        request(app)
+            .post("/")
+            .send({})
+            .set("Accept", "application/json")
+            .expect("Content-Type", /json/)
+            .expect(200, '{"status":"failed","errorMessage":"Request missing display name or username field!"}', done);
+    });
+});
+//# sourceMappingURL=attestation.test.js.map
