+++ package.json
@@ -63,7 +63,8 @@
     "express": "^4.16.4",
     "fido2-lib": "^2.1.1",
     "node-cache": "^4.2.0",
-    "string-to-arraybuffer": "^1.0.2"
+    "string-to-arraybuffer": "^1.0.2",
+    "url": "^0.11.0"
   },
   "nyc": {
     "extension": [
